add check for env
