+++ .env.example
@@ -16,3 +16,5 @@ PASSWORD_SECRET=xxxxxxxx
 
 EMAIL_USER=xxxxxxxx
 EMAIL_PASSWORD=xxxxxxxx
+
+ENVIRONMENT=dev
