--- constants/index.js
@@ -1,5 +1,0 @@
-module.exports = Object.freeze({
-  ENVIRONMENT: 'dev'
-});
-
-
