+++ src/Constants.js
@@ -1,0 +1,8 @@
+module.exports = Object.freeze({
+  env: {
+    dev: 'dev',
+    prod: 'prod'
+  }
+});
+
+
